Refactor repo structure
