+++ .github/workflows/docker-publish-server.yml
@@ -1,4 +1,4 @@
-name: Docker
+name: Docker Publish Server
 
 # This workflow uses actions that are not certified by GitHub.
 # They are provided by a third-party and are governed by
@@ -17,7 +17,7 @@ env:
   # Use docker.io for Docker Hub if empty
   REGISTRY: ghcr.io
   # github.repository as <account>/<repo>
-  IMAGE_NAME: ${{ github.repository }}/judge-server
+  IMAGE_NAME: ${{ github.repository }}/judger-server
 
 
 jobs:
@@ -66,12 +66,12 @@ jobs:
           images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
 
       # Build and push Docker image with Buildx (don't push on PR)
-      # https://github.com/docker/build-push-action
+      # https://github.com/docker/build-push-action     
       - name: Build and push Docker image
         id: build-and-push
         uses: docker/build-push-action@ac9327eae2b366085ac7f6a2d02df8aa8ead720a
         with:
-          file: docker/judge-server.dockerfile
+          file: docker/judger-server.dockerfile
           context: .
           push: ${{ github.event_name != 'pull_request' }}
           tags: ${{ steps.meta.outputs.tags }}
