+++ .github/workflows/rust-check.yml
@@ -1,4 +1,4 @@
-name: Test
+name: Rust Check
 
 on:
   push:
