+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["judge-core", "judge-service", "judge-server", "judge-cli"]
+members = ["judge-core", "judger"]
 resolver = "2"
