+++ docker/judger-server.dockerfile
@@ -1,17 +1,17 @@
 FROM rust:latest as build
 
 COPY judge-core /usr/src/judge-core
-COPY judge-server /usr/src/judge-server
-WORKDIR /usr/src/judge-server
+COPY judger /usr/src/judger
+WORKDIR /usr/src/judger
 
 RUN apt update && apt install -y libseccomp-dev gcc
-RUN cargo build --bin judge-server --release
+RUN cargo build --bin judger-server --release
 
 
 FROM ubuntu:latest
 
 RUN apt update && apt install -y libseccomp-dev gcc g++
-COPY --from=build /usr/src/judge-server/target/release/judge-server /usr/local/bin/judge-server
+COPY --from=build /usr/src/judger/target/release/judger-server /usr/local/bin/judger-server
 RUN mkdir /workspace
 WORKDIR /workspace
 COPY dev-problem-package /workspace/dev-problem-package
