--- judge-cli/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "judge-cli"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.0", features = ["derive"] }
-judge-core = { path = "../judge-core" }
-env_logger = "0.10.0"
-log = "0.4"
