--- judge-server/README.md
@@ -1,5 +1,0 @@
-### How to run
-`cargo run --bin judge-server -- --env-path ./judge-server/src/environment/.env.development`
-
-### How to visit OpenAPI
-visit `{HOST}/swagger-ui/`
