--- judge-server/postman/Judge Server.postman_collection.json
@@ -1,68 +1,0 @@
-{
-	"info": {
-		"_postman_id": "97809aca-56f3-4402-9d73-24c841f5c664",
-		"name": "Judge Server",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-		"_exporter_id": "12985580"
-	},
-	"item": [
-		{
-			"name": "v1/judge/hello-world",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"code\": \"#include <iostream>\\n\\nusing namespace std;\\n\\nint main() {\\n    string s;\\n    cin >> s;\\n    cout << \\\"Hello! \\\" << s << endl;\\n}\\n\",\r\n    \"language\": \"Cpp\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "{{Judger Host}}/api/v1/judge/hello-world",
-					"host": [
-						"{{Judger Host}}"
-					],
-					"path": [
-						"api",
-						"v1",
-						"judge",
-						"hello-world"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "v1/judge/hello-world TLE",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"code\": \"#include <time.h>\\r\\n#include <stdio.h>\\r\\n\\r\\nint main() {\\r\\n    clock_t start, step;\\r\\n    start = clock();\\r\\n    step = clock();\\r\\n    while (true) {\\r\\n        if (clock() - step > 1000000) {\\r\\n            printf(\\\"%lds has passed..\\\\n\\\", (clock() - start) \\/ 1000000);\\r\\n            step = clock();\\r\\n        }\\r\\n    }\\r\\n}\\r\\n\",\r\n    \"language\": \"Cpp\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "{{Judger Host}}/api/v1/judge/hello-world",
-					"host": [
-						"{{Judger Host}}"
-					],
-					"path": [
-						"api",
-						"v1",
-						"judge",
-						"hello-world"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
