--- judge-server/postman/OJ Lab Environment.postman_environment.json
@@ -1,21 +1,0 @@
-{
-	"id": "5fdf6ce1-e53b-44eb-b273-b44be3db8bab",
-	"name": "OJ Lab Environment",
-	"values": [
-		{
-			"key": "Judger Host",
-			"value": "localhost:8000",
-			"type": "default",
-			"enabled": true
-		},
-		{
-			"key": "Service Host",
-			"value": "localhost:8080",
-			"type": "default",
-			"enabled": true
-		}
-	],
-	"_postman_variable_scope": "environment",
-	"_postman_exported_at": "2023-09-03T02:21:12.497Z",
-	"_postman_exported_using": "Postman/10.17.4"
-}
