--- judge-service/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "judge-service"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-judge-core = { path = "../judge-core" }
-anyhow = "1.0"
-serde = "1"
-serde_derive = "1"
-serde_json = "1"
-log = "0.4"
-
-[dev-dependencies]
-env_logger = "0.10.0"
