--- judge-service/src/utils/mod.rs
@@ -1,1 +1,0 @@
-
