+++ judger/Cargo.toml
@@ -1,11 +1,16 @@
 [package]
-name = "judge-server"
+name = "judger"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+judge-core = { path = "../judge-core" }
+
+# CLI
+clap = { version = "4.0", features = ["derive"] }
+
 # Async runtime
 tokio = { version = "1", features = ["full"] }
 
@@ -37,4 +42,11 @@ thiserror = "1"
 
 uuid = { version = "1.4", features = ["serde", "v4"] }
 
-judge-core = { path = "../judge-core" }
+[[bin]]
+name ="judger-cli"
+path ="src/cli/main.rs"
+
+[[bin]]
+name ="judger-server"
+path ="src/server/main.rs"
+
