+++ judger/src/service/mod.rs
@@ -1,3 +1,2 @@
 pub mod error;
 pub mod package_manager;
-mod utils;
