+++ judger/tests/temp/judge-pd.json
@@ -1,0 +1,7 @@
+{
+  "test": {
+    "name": "test",
+    "revision": 1,
+    "package_type": "ICPC"
+  }
+}
